pyproject.toml
[tool.poetry]
name = "hnscraper"
version = "0.1.0"
description = ""
authors = ["ginruh"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.12"


aiohttp = "^3.9.5"
python-dotenv = "^1.0.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.31"}
asyncpg = "^0.29.0"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
listener = "src:listener.main"
top_stories = "src:top_stories.main"
